Refactored and working on updating to Python3
